Extract StartTimeRateLimiterAlgorithm
